refactor(db): extract pool init to db/src/pool.rs
