+++ db/src/lib.rs
@@ -9,6 +9,7 @@ pub mod author_photos;
 pub mod ebook;
 pub mod indexer;
 pub mod library_layout;
+pub mod pool;
 pub mod queries;
 pub mod scanner;
 pub mod thumbs;
@@ -17,5 +18,6 @@ pub mod worker;
 // Flatten the query layer so callers write `omnibus_db::list_books(...)`
 // instead of `omnibus_db::queries::list_books(...)`. Keeps callsites terse
 // and mirrors how `db.rs` looked before the extraction.
+pub use pool::*;
 pub use queries::*;
 pub use thumbs::{thumb_path_for, thumbs_dir, ThumbError, ThumbSize};
